requirements.txt
streamlit
requests
pandas
beautifulsoup4
plotly
python-dotenv
